refactor(ui): Manual content editor lives under the topology, not the source
The previous shape — `/datasources/<manualId>?topology=<topoId>` — was
structurally wrong. The data source page is meant for configuring the
*source itself* (name, type, upstream config); it should be
topology-agnostic. Observation content, by contrast, is per-
(topology, source) pair — so editing a Manual snapshot only makes
sense inside a topology context.

Restructure:

[New route]
- `/topologies/[id]/sources/[sourceId]/edit/+page.svelte`
  - YAML/JSON editor for any source 's observation in this topology
    (today only Manual is reachable through the UI, but the route
    is type-agnostic by design).
  - Reads via `latestSnapshot(topoId, sourceId)`, writes via
    `recordObservation(topoId, sourceId, graph)`.

[Reverted]
- `/datasources/[id]/+page.svelte`: removed the entire Manual-only
  YAML/JSON editor branch + its imports (YamlParser, Topology type,
  topologies store, graphToYaml, switchMode, handleSaveManual,
  parentTopologyId, manualTopology state). The page is back to
  doing one job — data source configuration.
- `api/datasources.ts`: removed the `GET /:id/topologies` helper I
  added earlier this turn; no longer needed since the editor lives
  under the topology namespace and never needs to derive its parent
  from a source id.

[Link updates]
- New-topology flow: `/topologies/+page.svelte` create modal now
  navigates to `/topologies/<id>/sources/<manualId>/edit` after the
  Manual attach.
- Topologies list row "Edit" button: same.
- `/topologies/[id]/settings` Actions card "Edit Manual" link: same.
- TopologySettings drawer Actions: same.

[Naming]
- The action label is now "Edit Manual content" (was "Edit Manual
  source") to disambiguate from "configure the source itself".

Bug fixed: arriving at the Manual editor without a `?topology=` query
showed "No parent topology context. Open this page from a topology 's
Sources tab…". That escape hatch is gone now — the route makes the
context structurally guaranteed.

Validation: bun run typecheck 34/34, svelte-check 0/0.
