+++ apps/server/web/src/routes/(app)/topologies/[id]/settings/+page.svelte
@@ -1198,11 +1198,11 @@
           <div class="card-body">
             {#if topology.manualSourceId}
               <a
-                href="/datasources/{topology.manualSourceId}?topology={topology.id}"
+                href="/topologies/{topology.id}/sources/{topology.manualSourceId}/edit"
                 class="btn btn-secondary w-full justify-center"
               >
                 <PencilSimpleIcon size={16} class="mr-2" />
-                Edit Manual source
+                Edit Manual content
               </a>
             {/if}
           </div>
